% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{check_design}
\alias{check_design}
\title{Internal check of whether inpu data frame (or tibble) is of the right format}
\usage{
check_design(X, method = "seq")
}
\arguments{
\item{method}{}
}
\description{
Do not call this function on its own
}
