% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{check_family}
\alias{check_family}
\title{Internal check of whether class of response is compatible with family}
\usage{
check_family(y, family)
}
\arguments{
\item{family}{}
}
\description{
Do not call this function on its own
}
