% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{eval_fdp}
\alias{eval_fdp}
\title{False discovery proportion (fdp) as function of selection and known negatives:}
\usage{
eval_fdp(selected, negatives)
}
\arguments{
\item{selected}{vector of indices of selected variables}

\item{negatives}{vector of indices of known null variables (that don't influence response)}
}
\value{
false discovery rate
}
\description{
False discovery proportion (fdp) as function of selection and known negatives:
}
\examples{
library(seqknockoff)

eval_fdp(selected=c(1,2,3,5), negatives=5:10)
}
