% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{generate_y}
\alias{generate_y}
\title{Simulate Gaussian response from a sparse regression model}
\usage{
generate_y(x, p_nn, a)
}
\arguments{
\item{x}{matrix corresponding to the regression design matrix}

\item{p_nn}{number of non-null covariate predictors. The regression coefficients corresponding to
columns 1:p_nn of x will be non-zero, all other are set to zero.}

\item{a}{amplitude of non-null regression coefficients}
}
\description{
Simulate Gaussian response from a sparse regression model
}
\examples{
library(seqknockoff)

set.seed(1)

# Simulate 4 Gaussian and 2 binary covariate predictors:
X <- generate_X(n=100, p=6, p_b=2, cov_type="cov_equi", rho=0.5)

# Obtain the corresponding model matrix:
x <- model.matrix(~., data=X)[,-1]

# Scale the binary dummy-variables so column-wise variance of x is same:
which.factor <- as.numeric(which(sapply(X, is.factor)))
x[,which.factor] <- x[,which.factor]/(0.5*sqrt(nrow(x)))

# Simulate response from model y = 2*x[,1] + 2*x[,2] + epsilon, where epsilon ~ N(0,1)
y <- generate_y(x, p_nn=2, a=2)
}
