% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simknockoffs.R
\name{knockoffs_mx}
\alias{knockoffs_mx}
\title{Gaussian MX-knockoffs for continuous variables}
\usage{
knockoffs_mx(X)
}
\arguments{
\item{X}{data.frame (or tibble) with "numeric" columns only. The number of columns, ncol(X) needs to be > 2.}
}
\value{
Second-order multivariate Gaussian knockoff copy of X
}
\description{
Gaussian MX-knockoffs for continuous variables
}
\details{
\code{knockoffs_mx} performs MX knockoff simulation.
}
\examples{
#' library(seqknockoff)

set.seed(1)

X <- generate_X(n=100, p=6, p_b=0, cov_type="cov_equi", rho=0.5)

Xk <- knockoffs_mx(X)
}
