% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockoff_filters.R
\name{multi_select}
\alias{multi_select}
\title{Select variables based on (heuristic) mode of multiple variable selections}
\usage{
multi_select(S, p, trim = 0.5)
}
\arguments{
\item{S}{list of variable selection indices}

\item{p}{number of variables. Each element of the list of selection indices should be a subset of 1:p.}

\item{trim}{trimming probability threshold. A sensible default is \code{trim=0.5}.}
}
\value{
a single "most frequent" variable selection among the multiple selections in S.
}
\description{
Select variables based on (heuristic) mode of multiple variable selections
}
\examples{
library(seqknockoff)

set.seed(1)

S <- list(c(sample(1:10,7), sample(11:20,1)),
c(sample(1:10,7), sample(11:20,1)),
c(sample(1:10,7), sample(11:20,1)))

multi_select(S, p=20)
}
