% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Heatmap of multiple variable selections ordered by importance}
\usage{
plot_heatmap(S, labels, nbcocluster = c(5, 5))
}
\arguments{
\item{S}{list of variable selection indices}

\item{labels}{labels of actual variable names corresponding to selection indices.}

\item{nbcocluster}{bivariate vector c(number of variable clusters, number of selection clusters).
The former number must be specified less than length(labels) and the latter must be less than length(S).}
}
\value{
plot of heatmap
}
\description{
Heatmap of multiple variable selections ordered by importance
}
\details{
The list of selected variables is converted to a length(labels)-by-length(S) binary
heatmap where each entry is either 1 if variable is selected and 0 otherwise. To help visualize most
important variables we perform coclustering of both selections and variables (with the blockcluster package).
}
\examples{
library(seqknockoff)

set.seed(1)

S <- list(c(sample(1:10,7), sample(11:20,1)),
c(sample(1:10,7), sample(11:20,1)),
c(sample(1:10,7), sample(11:20,1)))

plot_heatmap(S, labels=1:20, nbcocluster=c(5,3))
}
