% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_knockoff.R
\name{seqknockoff}
\alias{seqknockoff}
\title{Sequential knockoffs for continuous and categorical variables}
\usage{
seqknockoff(X, seq_simulator = sim_EN, ...)
}
\arguments{
\item{X}{data.frame with "numeric" and "factor" columns only (or tibble with <dbl> and <fct> columns). The number of columns, ncol(X) needs to be > 2.}

\item{seq_simulator}{function that simulates sequential knockoffs. Default is the function \code{sim_EN}, which simulates response from an estimated elastic-net model}

\item{...}{other parameters passed to the function seq_simulator. For the default (elastic-net sequential simulator, \code{seq_simulator = sim_EN})
these other parameters are passed to cv.glmnet.}
}
\value{
sequential knockoff copy of X. A data.frame or tibble of same type and dimensions as X.
}
\description{
Sequential knockoffs for continuous and categorical variables
}
\details{
\code{seqknockoff} performs sequential knockoff simulation.
}
\examples{
library(seqknockoff)

set.seed(1)

X <- generate_X(n=100, p=6, p_b=2, cov_type="cov_equi", rho=0.5)

# knockoffs based on sequential elastic-net regression with elastic-net penalty alpha:
Xk <- seqknockoff(X, alpha=0.5)
}
