% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_knockoff.R
\name{sim_EN}
\alias{sim_EN}
\title{Simulate from elastic-net regression model}
\usage{
sim_EN(y = y, x = x, ...)
}
\arguments{
\item{y}{response vector (either "numeric" or "factor") that gets passed to cv.glmnet}

\item{x}{input matrix that gets passed to cv.glmnet}

\item{...}{other parameters passed to the function cv.glmnet}
}
\description{
Simulate from elastic-net regression model
}
\examples{
library(seqknockoff)

set.seed(1)

x = matrix(rnorm(100 * 20), 100, 20)
y = x[,1] + rnorm(100)

# simulate from elastic-net regression with elastic-net penalty alpha:
ysim = sim_EN(y=y, x=x, alpha=0.5)

# simulated versus input response:
plot(y, ysim)
}
