% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\name{generate_y}
\alias{generate_y}
\title{Simulate Gaussian response from a sparse regression model}
\usage{
generate_y(X, p_nn, a)
}
\arguments{
\item{X}{matrix corresponding to the regression design matrix.
Numeric columns of X should have variance = 1/nrow(X), default behavior of generate_X.}

\item{p_nn}{number of non-null covariate predictors.
The regression coefficients (beta) corresponding to
columns 1:p_nn of x will be non-zero, all other are set to zero.}

\item{a}{amplitude of non-null regression coefficients}
}
\value{
simulated Gaussian response from regression model y = X%*%beta+epsilon, where epsilon ~ N(0,1). Binary factor columns of X are
transformed to dummy 0/1 variables and then scaled so variance = 1/nrow(X).
}
\description{
Simulate Gaussian response from a sparse regression model
}
\examples{
library(seqknockoff)

set.seed(1)

# Simulate 4 Gaussian and 2 binary covariate predictors:
X <- generate_X(n=100, p=6, p_b=2, cov_type="cov_equi", rho=0.5)

# Simulate response from model y = 2*X[,1] + 2*X[,2] + epsilon, where epsilon ~ N(0,1)
y <- generate_y(X, p_nn=2, a=2)
}
